Some improvements in code, login api integration finished
