+++ src/helpers/api/index.js
@@ -51,9 +51,22 @@ const getUserData = userId =>
          .catch(err => reject(err.response.data));
    });
 
+const loginUser = ({ email, password }) =>
+   new Promise((resolve, reject) => {
+      axios
+         .post("/login-user", { email, password })
+         .then(response => {
+            const resData = response.data;
+            if (!resData.success) reject(resData);
+            else resolve(resData);
+         })
+         .catch(err => reject(err.response.data));
+   });
+
 export {
    sendConfirmationEmail,
    confirmAccount,
    updateUserProfile,
    getUserData,
+   loginUser,
 };
