+++ src/pages/SignUp.jsx
@@ -11,7 +11,7 @@ function SignUp(props) {
 
    useEffect(() => {
       let userAuth = JWT.decode(getUserAuth());
-      if (userAuth?.active) return history.push("/");
+      if (userAuth?.active) return history.goBack();
    }, []);
 
    return (
